+++ packages/schematics/angular/enum/index.ts
@@ -6,46 +6,12 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import { strings } from '@angular-devkit/core';
-import {
-  Rule,
-  Tree,
-  apply,
-  applyTemplates,
-  chain,
-  mergeWith,
-  move,
-  noop,
-  url,
-} from '@angular-devkit/schematics';
-import { applyLintFix } from '../utility/lint-fix';
-import { parseName } from '../utility/parse-name';
-import { createDefaultPath } from '../utility/workspace';
+import { Rule } from '@angular-devkit/schematics';
+import { generateFromFiles } from '../utility/generate-from-files';
 import { Schema as EnumOptions } from './schema';
 
 export default function (options: EnumOptions): Rule {
-  return async (host: Tree) => {
-    if (options.path === undefined) {
-      options.path = await createDefaultPath(host, options.project as string);
-    }
+  options.type = options.type ? `.${options.type}` : '';
 
-    const parsedPath = parseName(options.path, options.name);
-    options.name = parsedPath.name;
-    options.path = parsedPath.path;
-
-    options.type = options.type ? `.${options.type}` : '';
-
-    const templateSource = apply(url('./files'), [
-      applyTemplates({
-        ...strings,
-        ...options,
-      }),
-      move(parsedPath.path),
-    ]);
-
-    return chain([
-      mergeWith(templateSource),
-      options.lintFix ? applyLintFix(options.path) : noop(),
-    ]);
-  };
+  return generateFromFiles(options);
 }
