+++ packages/schematics/angular/guard/index.ts
@@ -6,72 +6,37 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import { strings } from '@angular-devkit/core';
-import {
-  Rule,
-  SchematicsException,
-  Tree,
-  apply,
-  applyTemplates,
-  chain,
-  filter,
-  mergeWith,
-  move,
-  noop,
-  url,
-} from '@angular-devkit/schematics';
-import { applyLintFix } from '../utility/lint-fix';
-import { parseName } from '../utility/parse-name';
-import { createDefaultPath } from '../utility/workspace';
+import { Rule, SchematicsException } from '@angular-devkit/schematics';
+import { generateFromFiles } from '../utility/generate-from-files';
 import { Implement as GuardInterface, Schema as GuardOptions } from './schema';
 
 export default function (options: GuardOptions): Rule {
-  return async (host: Tree) => {
-    if (options.path === undefined) {
-      options.path = await createDefaultPath(host, options.project as string);
-    }
+  if (!options.implements) {
+    throw new SchematicsException('Option "implements" is required.');
+  }
 
-    if (!options.implements) {
-      throw new SchematicsException('Option "implements" is required.');
-    }
+  const implementations = options.implements
+    .map((implement) => (implement === 'CanDeactivate' ? 'CanDeactivate<unknown>' : implement))
+    .join(', ');
+  const commonRouterNameImports = ['ActivatedRouteSnapshot', 'RouterStateSnapshot'];
+  const routerNamedImports: string[] = [...options.implements, 'UrlTree'];
 
-    const implementations = options.implements
-      .map((implement) => (implement === 'CanDeactivate' ? 'CanDeactivate<unknown>' : implement))
-      .join(', ');
-    const commonRouterNameImports = ['ActivatedRouteSnapshot', 'RouterStateSnapshot'];
-    const routerNamedImports: string[] = [...options.implements, 'UrlTree'];
+  if (options.implements.includes(GuardInterface.CanLoad)) {
+    routerNamedImports.push('Route', 'UrlSegment');
 
-    if (options.implements.includes(GuardInterface.CanLoad)) {
-      routerNamedImports.push('Route', 'UrlSegment');
-
-      if (options.implements.length > 1) {
-        routerNamedImports.push(...commonRouterNameImports);
-      }
-    } else {
+    if (options.implements.length > 1) {
       routerNamedImports.push(...commonRouterNameImports);
     }
+  } else {
+    routerNamedImports.push(...commonRouterNameImports);
+  }
 
-    routerNamedImports.sort();
+  routerNamedImports.sort();
 
-    const implementationImports = routerNamedImports.join(', ');
-    const parsedPath = parseName(options.path, options.name);
-    options.name = parsedPath.name;
-    options.path = parsedPath.path;
+  const implementationImports = routerNamedImports.join(', ');
 
-    const templateSource = apply(url('./files'), [
-      options.skipTests ? filter((path) => !path.endsWith('.spec.ts.template')) : noop(),
-      applyTemplates({
-        implementations,
-        implementationImports,
-        ...strings,
-        ...options,
-      }),
-      move(parsedPath.path + (options.flat ? '' : '/' + strings.dasherize(options.name))),
-    ]);
-
-    return chain([
-      mergeWith(templateSource),
-      options.lintFix ? applyLintFix(options.path) : noop(),
-    ]);
-  };
+  return generateFromFiles(options, {
+    implementations,
+    implementationImports,
+  });
 }
