+++ packages/schematics/angular/interceptor/index.ts
@@ -6,47 +6,16 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import { strings } from '@angular-devkit/core';
-import {
-  Rule,
-  Tree,
-  apply,
-  applyTemplates,
-  chain,
-  filter,
-  mergeWith,
-  move,
-  noop,
-  url,
-} from '@angular-devkit/schematics';
-import { applyLintFix } from '../utility/lint-fix';
-import { parseName } from '../utility/parse-name';
-import { createDefaultPath } from '../utility/workspace';
+import { Rule } from '@angular-devkit/schematics';
+import { generateFromFiles } from '../utility/generate-from-files';
 import { Schema as InterceptorOptions } from './schema';
 
 export default function (options: InterceptorOptions): Rule {
-  return async (host: Tree) => {
-    if (options.path === undefined) {
-      options.path = await createDefaultPath(host, options.project as string);
-    }
+  // This schematic uses an older method to implement the flat option
+  const flat = options.flat;
+  options.flat = true;
 
-    const parsedPath = parseName(options.path, options.name);
-    options.name = parsedPath.name;
-    options.path = parsedPath.path;
-
-    const templateSource = apply(url('./files'), [
-      options.skipTests ? filter((path) => !path.endsWith('.spec.ts.template')) : noop(),
-      applyTemplates({
-        ...strings,
-        'if-flat': (s: string) => (options.flat ? '' : s),
-        ...options,
-      }),
-      move(parsedPath.path),
-    ]);
-
-    return chain([
-      mergeWith(templateSource),
-      options.lintFix ? applyLintFix(options.path) : noop(),
-    ]);
-  };
+  return generateFromFiles(options, {
+    'if-flat': (s: string) => (flat ? '' : s),
+  });
 }
