+++ packages/schematics/angular/resolver/index.ts
@@ -6,42 +6,10 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import { strings } from '@angular-devkit/core';
-import {
-  Rule,
-  Tree,
-  apply,
-  applyTemplates,
-  chain,
-  filter,
-  mergeWith,
-  move,
-  noop,
-  url,
-} from '@angular-devkit/schematics';
-import { parseName } from '../utility/parse-name';
-import { createDefaultPath } from '../utility/workspace';
+import { Rule } from '@angular-devkit/schematics';
+import { generateFromFiles } from '../utility/generate-from-files';
 import { Schema } from './schema';
 
 export default function (options: Schema): Rule {
-  return async (host: Tree) => {
-    if (options.path === undefined) {
-      options.path = await createDefaultPath(host, options.project as string);
-    }
-
-    const parsedPath = parseName(options.path, options.name);
-    options.name = parsedPath.name;
-    options.path = parsedPath.path;
-
-    const templateSource = apply(url('./files'), [
-      options.skipTests ? filter((path) => !path.endsWith('.spec.ts.template')) : noop(),
-      applyTemplates({
-        ...strings,
-        ...options,
-      }),
-      move(parsedPath.path + (options.flat ? '' : '/' + strings.dasherize(options.name))),
-    ]);
-
-    return chain([mergeWith(templateSource)]);
-  };
+  return generateFromFiles(options);
 }
